Rebuilt code to access ePub ISBN based on file to bypass .DS_Store and only pick up files starting with 978 which match all ePubs to date. If it continues to be an issue we can target 'endswith' epub.
Mistake found on a recent update to the dir_list, initial value was set as 1 which caused errors. Set to 0.

EOD FRI: Need to handel ln 215 with "open" cmd, rather than assignment to avoid FileNotFoundError when the excess music dir isn't present. As it stands now it works but it isn't quite elegant.

Fixed issue and runs without error regardless if extra music dir exists.